extract sharable components
